falling back to the root calss
the root pseudo class could work better, in case the HTML is loaded inside of the iframe or some screen reader.

Also all elements can access the variables in the root class, not just the HTML, as previously set.

Let's see what works best. the HTML will still have the dark CSS class to change it's colors.

in any way root is most of the time same as HTML, if not loaded inside of an iFrame.